requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
apscheduler==3.6.3
    # via python-telegram-bot
backports-zoneinfo==0.2.1
    # via tzlocal
cachetools==4.2.2
    # via python-telegram-bot
certifi==2021.10.8
    # via
    #   cloudinary
    #   python-telegram-bot
    #   requests
charset-normalizer==2.0.12
    # via requests
cloudinary==1.29.0
    # via -r requirements.in
faunadb==4.2.0
    # via -r requirements.in
future==0.18.2
    # via faunadb
h2==2.6.2
    # via hyper
hpack==3.0.0
    # via h2
hyper==0.7.0
    # via faunadb
hyperframe==3.2.0
    # via
    #   h2
    #   hyper
idna==3.3
    # via requests
iso8601==1.0.2
    # via faunadb
python-dotenv==0.20.0
    # via -r requirements.in
python-http-client==3.3.7
    # via sendgrid
python-telegram-bot==13.11
    # via -r requirements.in
pytz==2022.1
    # via
    #   apscheduler
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.27.1
    # via faunadb
sendgrid==6.9.7
    # via -r requirements.in
six==1.16.0
    # via
    #   apscheduler
    #   cloudinary
starkbank-ecdsa==2.0.3
    # via sendgrid
tornado==6.1
    # via python-telegram-bot
tzdata==2022.1
    # via tzlocal
tzlocal==4.2
    # via apscheduler
urllib3==1.26.9
    # via
    #   cloudinary
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
